# backend/API-definition.yaml
openapi: 3.0.3
info:
  title: API de Gestión de Empleados, Fichajes, Vacaciones y Dashboard
  version: 1.0.0
  description: |
    Este documento describe la definición de la API para la gestión de autenticación con Google + JWT, 
    CRUD de empleados, control de fichajes, gestión de vacaciones y rutas de dashboard/estadísticas.
servers:
  - url: /
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Empleado:
      type: object
      properties:
        id:
          type: integer
          example: 5
        nombre:
          type: string
          example: "Ana López"
        email:
          type: string
          format: email
          example: "ana.lopez@example.com"
        cargo:
          type: string
          example: "Desarrolladora"
        fecha_contratacion:
          type: string
          format: date-time
          example: "2025-06-15T00:00:00Z"
        supervisor_id:
          type: integer
          nullable: true
          example: 2
        rol:
          type: string
          enum:
            - empleado
            - supervisor
          example: "empleado"
      required:
        - id
        - nombre
        - email
        - cargo
        - fecha_contratacion
        - rol
...
    EmpleadoCreateRequest:
      type: object
      properties:
        nombre:
          type: string
          example: "Ana López"
        email:
          type: string
          format: email
          example: "ana.lopez@example.com"
        cargo:
          type: string
          example: "Desarrolladora"
        fecha_contratacion:
          type: string
          format: date
          example: "2025-06-15"
        supervisor_id:
          type: integer
          nullable: true
          example: 2
        rol:
          type: string
          enum:
            - empleado
            - supervisor
          example: "empleado"
      required:
        - nombre
        - email
        - cargo
        - fecha_contratacion
        - rol
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Descripción del error"
      required:
        - message
...
paths:
...
  /empleados/:
    post:
      summary: Crear nuevo empleado/supervisor
      description: |
        Crea un nuevo empleado o supervisor. Solo rol=supervisor puede invocar.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpleadoCreateRequest'
      responses:
        '201':
          description: Empleado creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empleado'
        '400':
          description: JSON mal formado o falta campo obligatorio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token faltante o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Token no pertenece a un supervisor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email ya existe en BD
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Listar empleados asignados
      description: |
        Si rol=supervisor: devuelve empleados cuyo supervisor_id coincide con usuario_id y también su propio registro.
        Si rol=empleado: devuelve únicamente su propio registro.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de empleados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Empleado'
        '401':
          description: Token faltante o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...